refactor: create users file
